+++ liquid-interpreter/src/lib.rs
@@ -23,7 +23,7 @@ mod context;
 mod expression;
 mod filter;
 mod filter_chain;
-mod globals;
+mod store;
 mod registry;
 mod renderable;
 mod template;
@@ -34,7 +34,7 @@ pub use self::context::*;
 pub use self::expression::*;
 pub use self::filter::*;
 pub use self::filter_chain::*;
-pub use self::globals::*;
+pub use self::store::*;
 pub use self::registry::*;
 pub use self::renderable::*;
 pub use self::template::*;
